+++ README.md
@@ -1,1 +1,38 @@
-# App
+
+# Dental tooth fairy website
+
+We remove tooth without asking. Free of cost.
+
+This project uses nodejs, passportjs, css, tailwindcss, node mailer, postgress db, etc, under the hood.
+
+Essentially, this is a demo website We've create to flex our might in this ocean of noob developers. Unless you're are noob yourself, don't even think about cloning and stealing out website. 
+
+This is a collective initiative by Ayeesha and Donald duck.
+
+If you are thinking of cloning or forking this project please consider supporting my hard work by give a start to this repo. 
+
+The website is deployed [here](wait_for_deployment).
+
+
+## Badges
+
+[![MIT License](https://img.shields.io/badge/License-MIT-green.svg)](https://choosealicense.com/licenses/mit/) 
+[![GPLv3 License](https://img.shields.io/badge/License-GPL%20v3-yellow.svg)](https://opensource.org/licenses/)
+[![AGPL License](https://img.shields.io/badge/license-AGPL-blue.svg)](http://www.gnu.org/licenses/agpl-3.0)
+
+
+## Acknowledgements
+
+ - Angella yu
+ - Our forged teeth
+
+
+
+## Authors
+- [@ayeesha](https://www.github.com/ayeesha)
+- [@sonukuldeep](https://www.github.com/sonukuldeep)
+
+
+## 🛠 Skills
+
+[![My Skills](https://skillicons.dev/icons?i=js,ts,html,css,tailwind,sass,nodejs,react,vue,flask,rust,python,php,solidity,mongodb,mysql,prisma,figma)](https://github.com/donald-gg)
