--- app/views/footer.ejs
@@ -1,3 +1,0 @@
-<footer>
-    footer goes here
-</footer>
