--- app/views/header.ejs
@@ -1,3 +1,0 @@
-<div class="hero">
-    Hero section goes here
-</div>
