--- app/views/main.ejs
@@ -1,5 +1,0 @@
-<main>
-    Main content goes here
-
-    as project grows main will be subdivided to different sections
-</main>
