--- app/views/navbar.ejs
@@ -1,3 +1,0 @@
-<nav>
-    navbar goes here
-</nav>
