+++ package.json
@@ -1,0 +1,19 @@
+{
+    "name": "dental-website",
+    "version": "1.0.0",
+    "description": "A boilerplate for express website",
+    "main": "index.js",
+    "scripts": {
+      "dev": "nodemon ./index.js"
+    },
+    "author": "donald and",
+    "license": "MIT",
+    "dependencies": {
+      "dotenv": "^16.0.3",
+      "ejs": "^3.1.9",
+      "express": "^4.18.2"
+    },
+    "devDependencies": {
+      "nodemon": "^2.0.22"
+    }
+  }
