+++ views/footer.ejs
@@ -1,0 +1,3 @@
+<footer>
+    footer goes here
+</footer>
