+++ views/head.ejs
@@ -1,0 +1,17 @@
+<meta charset="UTF-8">
+<meta http-equiv="X-UA-Compatible" content="IE=edge">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
+<meta name="viewport" content="width=device-width, initial-scale=1.0" />
+<meta name="description" content="Checkout my tooth fairy website!" />
+<meta name="keywords" content="dental clinic webiste, dental clinic, free checkup" />
+<meta name="author" content="tooth fairy" />
+<meta property="og:image" content="/og_image.jpg" />
+<meta property="og:url" content="toothfairywebsite.com" />
+<meta property="og:type" content="website" />
+<meta property="og:site_name" content="toothfairywebsite.com" />
+<meta property="og:description" content="Get the best dental job ever. We remove teeth without asking! Absolutely free" />
+<meta property="og:locale" content="en_US" />
+<link rel="stylesheet" href="/assets/css/style.css">
+<script src="/assets/js/script.js" defer></script>
+<title>Document</title>
