+++ views/header.ejs
@@ -1,0 +1,3 @@
+<div class="hero">
+    Hero section goes here
+</div>
