+++ views/index.ejs
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <%- include('head') %>
+</head>
+
+<body>
+    <%- include('navbar') %>
+    <%- include('header') %>
+    <%- include('main') %>
+    <%- include('footer') %>
+</body>
+
+</html>
