+++ views/main.ejs
@@ -1,0 +1,5 @@
+<main>
+    Main content goes here
+
+    as project grows main will be subdivided to different sections
+</main>
