+++ views/navbar.ejs
@@ -1,0 +1,3 @@
+<nav>
+    navbar goes here
+</nav>
